[TASK] Add option to include TypoScript from custom path
Currently, it is not possible to use a custom template path.
This can be useful if you have plenty of sites which do not use
a dedicated extension for each site.

This change adds the site setting "typoScriptTemplatePath" which
allows to define a custom path for the TypoScript templates
to be included.